// src/app/components/supplier/model/supplierScoreCardSummaryModel.ts
export class supplierScoreCardSummaryModel{
    supplierName: string;
    tradelicenseno: string;
    submissionClassification: string;
    submittedDate: string;
    score: string;
    status: string;
    noteToSupplier: string;
    givenScore: string;
    totalScore: string;
    qualStatusDropdownValue: string;
    remarks: string;
    parentCategory: string;
    childCategory: string;
    subCategory: string;
}